Introduce packages to separate domain from data and presentation
