+++ ModelViewIntentSample/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
             android:supportsRtl="true"
             android:theme="@style/AppTheme"
             android:name=".MoviesApplication">
-        <activity android:name=".search.SearchActivity">
+        <activity android:name=".search.presentation.SearchActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
