+++ ModelViewIntentSample/search/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
             android:theme="@style/AppTheme">
-        <activity android:name=".SearchActivity"
+        <activity android:name=".presentation.SearchActivity"
                   android:screenOrientation="portrait"/>
     </application>
