+++ ModelViewIntentSample/search/src/main/java/com/novoda/movies/mvi/search/data/ApiSearchResults.kt
@@ -1,4 +1,4 @@
-package com.novoda.movies.mvi.search.api
+package com.novoda.movies.mvi.search.data
 
 
 internal data class ApiSearchResults(
