+++ ModelViewIntentSample/search/src/main/java/com/novoda/movies/mvi/search/domain/SearchResultsModel.kt
@@ -1,4 +1,4 @@
-package com.novoda.movies.mvi.search
+package com.novoda.movies.mvi.search.domain
 
 import io.reactivex.Observable
