+++ ModelViewIntentSample/search/src/main/java/com/novoda/movies/mvi/search/presentation/SearchResultsViewable.kt
@@ -1,4 +1,4 @@
-package com.novoda.movies.mvi.search.view
+package com.novoda.movies.mvi.search.presentation
 
 internal interface SearchResultsViewable {
     fun showResults(results: ViewSearchResults)
