+++ ModelViewIntentSample/search/src/main/java/com/novoda/movies/mvi/search/presentation/ViewSearchResults.kt
@@ -1,4 +1,4 @@
-package com.novoda.movies.mvi.search.view
+package com.novoda.movies.mvi.search.presentation
 
 import java.net.URL
